refactor: 기물 점수 계산 로직을 클래스로 추출
[- 책임 분리를 위해 기물 점수 계산 로직을 기물 점수 계산 클래스로 추출]
