+++ src/main/java/com/wootecam/chess/board/Board.java
@@ -14,9 +14,13 @@ public class Board {
     public static final int MAX_ROW = 8;
 
     private final Rank[] ranks;
+    private final ScoreCalculationRule scoreCalculationRule;
+
     private int totalPieces;
 
-    public Board() {
+    public Board(ScoreCalculationRule scoreCalculationRule) {
+        this.scoreCalculationRule = scoreCalculationRule;
+
         this.ranks = new Rank[MAX_ROW];
         for (int i = 0; i < MAX_ROW; i++) {
             this.ranks[i] = new Rank();
@@ -66,35 +70,7 @@ public class Board {
     }
 
     public double calculateScore(Color color) {
-        double score = Arrays.stream(ranks)
-                .mapToDouble(r -> r.calculateScoreExceptPawn(color))
-                .sum();
-        score += calculatePawnScore(color);
-
-        return score;
-    }
-
-    private double calculatePawnScore(Color color) {
-        double score = 0;
-
-        for (int col = 0; col < MAX_COL; ++col) {
-            int count = countPawnsInColumn(color, col);
-
-            if (count == 1) {
-                score += PieceType.PAWN.point;
-            } else if (count >= 2) {
-                score += PieceType.PAWN.point * count / 2;
-            }
-        }
-
-        return score;
-    }
-
-    private int countPawnsInColumn(Color color, int col) {
-        return (int) Arrays.stream(ranks)
-                .map(r -> r.get(col))
-                .filter(p -> p.isColor(color) && p.isPawn())
-                .count();
+        return scoreCalculationRule.apply(ranks, color);
     }
 
     public List<Piece> getPiecesSortedByScore(Color color, Order order) {
@@ -105,71 +81,5 @@ public class Board {
                     return order.isAsc() ? compare : -compare;
                 })
                 .toList();
-    }
-
-    private static class Rank {
-        private final Piece[] squares;
-
-        public Rank() {
-            this.squares = new Piece[MAX_COL];
-            Arrays.fill(squares, Piece.BLANK);
-        }
-
-        public void place(Piece piece, int index) {
-            checkValidIndex(index);
-            checkEmptySquare(index);
-
-            squares[index] = piece;
-        }
-
-        public void clearSquare(int index) {
-            checkValidIndex(index);
-
-            squares[index] = Piece.BLANK;
-        }
-
-        private void checkValidIndex(int index) {
-            if (index < 0 || index >= MAX_COL) {
-                throw new IllegalArgumentException("The index is invalid: " + index);
-            }
-        }
-
-        private void checkEmptySquare(int index) {
-            if (squares[index].isPiece()) {
-                throw new IllegalArgumentException("The square is already occupied");
-            }
-        }
-
-        public String print() {
-            return Arrays.stream(squares)
-                    .map(Piece::getRepresentation)
-                    .map(r -> r.value)
-                    .collect(Collectors.joining());
-        }
-
-        public Piece get(int index) {
-            checkValidIndex(index);
-
-            return squares[index];
-        }
-
-        public int countPiece(PieceType type, Color color) {
-            return (int) Arrays.stream(squares)
-                    .filter(piece -> piece.hasTypeAndColor(type, color))
-                    .count();
-        }
-
-        public double calculateScoreExceptPawn(Color color) {
-            return Arrays.stream(squares)
-                    .filter(p -> p.isColor(color) && !p.isPawn())
-                    .mapToDouble(p -> p.getType().point)
-                    .sum();
-        }
-
-        public List<Piece> getPieces(Color color) {
-            return Arrays.stream(squares)
-                    .filter(p -> p.isColor(color))
-                    .toList();
-        }
     }
 }
