+++ src/main/java/com/wootecam/chess/game/ChessGameManager.java
@@ -1,6 +1,7 @@
 package com.wootecam.chess.game;
 
 import com.wootecam.chess.board.BoardInitializer;
+import com.wootecam.chess.board.ScoreCalculationRule;
 import com.wootecam.chess.common.Reader;
 
 public class ChessGameManager {
@@ -9,6 +10,7 @@ public class ChessGameManager {
     private static final String CMD_MOVE = "move";
 
     private static final BoardInitializer boardInitializer = new BoardInitializer();
+    private static final ScoreCalculationRule scoreCalculationRule = new ScoreCalculationRule();
 
     private final Reader reader;
 
@@ -55,7 +57,7 @@ public class ChessGameManager {
     }
 
     private ChessGame startGame() {
-        ChessGame chessGame = new ChessGame(boardInitializer);
+        ChessGame chessGame = new ChessGame(boardInitializer, scoreCalculationRule);
         chessGame.start();
 
         return chessGame;
